# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de BarberMusic&Spa",
        "description": "Documentación para la API utilizada en el sistema de BarberMusic&Spa. Aquí puedes probar los endpoints de autenticación.",
        "contact": {
            "email": "soporte@barbermusic.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/Catalog_servicios/servicios": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Obtener todos los servicios",
                "operationId": "d28910473a5737d9617db146885ce6b8",
                "responses": {
                    "200": {
                        "description": "Lista de servicios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "imagen_path": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "precio_base": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "duracion_minutos": {
                                                "type": "integer"
                                            },
                                            "categoria_id": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "especialidad_requerida_id": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "activo": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Crear un nuevo servicio",
                "operationId": "c8c41a565510cc9db15804997c48899a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Corte de Cabello Clásico"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Corte con máquina y tijera."
                                    },
                                    "precio_base": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250
                                    },
                                    "duracion_minutos": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "especialidad_requerida_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servicio creado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Servicio creado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Catalog_servicios/servicios/{id}": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Obtener un servicio por ID",
                "operationId": "dd2aa29302bcc75389920b2b8f680fa2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del servicio"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Actualizar un servicio",
                "operationId": "c9462cd5c3ea7a9111f046007776a128",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Corte de Barba"
                                    },
                                    "precio_base": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Servicio actualizado"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Eliminar un servicio (Soft Delete)",
                "operationId": "b46406bf33d1a6b624205c9ea3aea331",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Servicio eliminado"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token de acceso (JWT o Sanctum) sin la palabra 'Bearer'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}